backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
openpyxl
pypdf
openai
python-dotenv
sqlalchemy
psycopg2-binary
google-generativeai
